DoctorClinic Dao and Sercice

+++ Przychodnia/src/main/java/com/ai/przychodnia/dao/RegNotificationDao.java
@@ -19,5 +19,5 @@ public interface RegNotificationDao
 	
 	void deleteNotification(Reg_notification notification);
 	
-	int countNewNotifications();
+	long countNewNotifications();
 }
